+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to Movie Sync project.
 
 - Downloaded items are now moved to the downloaded queue only after they are truly downloaded in the torrent client (instead of when they are added)
 - Merged  `download_directory` from `config.json` and `qbittorrent.save_path` from `filelist_config.json`, now only `download_directory` is used
+- Fixed torrent deletion from qbittorrent
 
 ## [0.1.0] - 2025-10-21